requirements.txt
yt-dlp>=2023.12.30
youtube-transcript-api>=0.6.2
openai>=1.3.0
python-dotenv>=1.0.0
flask>=2.3.0
elevenlabs>=0.2.0
requests>=2.25.0
supabase>=2.3.0
postgrest>=0.16.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3